# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "这是一个用 Go 编写的博客后端 API 示例，支持用户、博客、评论、点赞等功能。",
        "title": "博客后端 API 文档",
        "contact": {
            "name": "API Support",
            "email": "youremail@example.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/comment": {
            "post": {
                "description": "创建新的评论",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "评论"
                ],
                "summary": "创建评论",
                "parameters": [
                    {
                        "description": "评论内容",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateCommentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Comment"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除指定评论",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "评论"
                ],
                "summary": "删除评论",
                "parameters": [
                    {
                        "description": "评论ID",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DeleteCommentRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "删除成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "未找到",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/like": {
            "post": {
                "description": "给博客点赞",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "点赞"
                ],
                "summary": "点赞",
                "parameters": [
                    {
                        "description": "点赞内容",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AddLikeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Like"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "取消对博客的点赞",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "点赞"
                ],
                "summary": "取消点赞",
                "parameters": [
                    {
                        "description": "点赞内容",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RemoveLikeRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "取消成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "未找到",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.AddLikeRequest": {
            "type": "object",
            "properties": {
                "blog_id": {
                    "type": "string"
                }
            }
        },
...
        "models.Comment": {
            "type": "object",
            "properties": {
                "blog_id": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "models.CreateCommentRequest": {
            "type": "object",
            "properties": {
                "blog_id": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                }
            }
        },
        "models.DeleteCommentRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "models.Like": {
            "type": "object",
            "properties": {
                "blog_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.RemoveLikeRequest": {
            "type": "object",
            "properties": {
                "blog_id": {
                    "type": "string"
                }
            }
        },
...
    }
}