# models/comment.go
package models

import (
	"time"

	"go.mongodb.org/mongo-driver/bson/primitive"
)

// Comment 评论结构
type Comment struct {
	ID        primitive.ObjectID `bson:"_id,omitempty" json:"id,omitempty"`
	BlogID    primitive.ObjectID `bson:"blog_id" json:"blog_id"`
	UserID    primitive.ObjectID `bson:"user_id" json:"user_id"`
	Username  string             `bson:"username" json:"username"`
	Content   string             `bson:"content" json:"content"`
	CreatedAt time.Time          `bson:"created_at" json:"created_at"`
	UpdatedAt time.Time          `bson:"updated_at" json:"updated_at"`
}

// CreateCommentRequest 创建评论请求
type CreateCommentRequest struct {
	BlogID  string `json:"blog_id"`
	Content string `json:"content"`
}

// DeleteCommentRequest 删除评论请求
type DeleteCommentRequest struct {
	ID string `json:"id"`
}